requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitarray==3.8.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
ckzg==2.1.7
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
cytoolz==1.1.0
datasets==4.7.0
defusedxml==0.7.1
dill==0.4.0
dnspython==2.8.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
filelock==3.20.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2025.12.0
ftfy==6.3.1
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
ipwhois==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lightgbm==4.6.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.3.5
packaging==26.0
pandas==3.0.1
parsimonious==0.10.0
pillow==12.0.0
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-whois==0.9.6
pytz==2026.1.post1
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
rich==14.3.3
rlp==4.1.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tldextract==5.3.1
tokenizers==0.22.2
toolz==1.1.0
torch==2.10.0+cpu
torch-geometric==2.7.0
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
web3==7.14.1
websockets==15.0.1
Werkzeug==3.1.6
win32_setctime==1.2.0
xxhash==3.6.0
yarl==1.23.0
